\name{rm.all}
\alias{rm.all}
\title{Remove all objects from the global environment}
\description{Removes all objects from the global environment, optionally also removing hidden objects.}
\usage{rm.all(include.hidden = FALSE)}
\arguments{
  \item{include.hidden}{ logical: should hidden objects also be removed?}
}
\author{Alastair Droop <alastair.droop@gmail.com>}
\seealso{\code{\link{clean}}}
\examples{rm.all()}
